requirements.txt
flask
flask-jwt-simple
flask-cors
psycopg2
sphinx
sqlalchemy
python-dateutil
werkzeug
python-dateutil
alembic
